Move NAM-related structs and functions to nam.cpp/nam.hpp
